Improve test related code
