+++ tests/package.json
@@ -30,9 +30,15 @@
     "testAddTheme": "mocha --timeout 9999999 -r ts-node/register './src/addTheme.test.ts'",
     "testCreateBase": "mocha --timeout 9999999 -r ts-node/register './src/createBase.test.ts'",
     "testCreateCollection": "mocha --timeout 9999999 -r ts-node/register './src/createCollection.test.ts'",
+    "testDeleteCollection": "mocha --timeout 9999999 -r ts-node/register './src/deleteCollection.test.ts'",
+    "testChangeCollectionIssuer": "mocha --timeout 9999999 -r ts-node/register './src/changeCollectionIssuer.test.ts'",
+    "testLockCollection": "mocha --timeout 9999999 -r ts-node/register './src/lockCollection.test.ts'",
     "testMintNft": "mocha --timeout 9999999 -r ts-node/register './src/mintNft.test.ts'",
+    "testBurnNft": "mocha --timeout 9999999 -r ts-node/register './src/burnNft.test.ts'",
     "testGetOwnedNfts": "mocha --timeout 9999999 -r ts-node/register './src/getOwnedNfts.test.ts'",
     "testSetNftProperty": "mocha --timeout 9999999 -r ts-node/register './src/setNftProperty.test.ts'",
+    "testSetCollectionProperty": "mocha --timeout 9999999 -r ts-node/register './src/setCollectionProperty.test.ts'",
+    "testAddNftResource": "mocha --timeout 9999999 -r ts-node/register './src/addNftResource.test.ts'",
     "testSetResourcePriorities": "mocha --timeout 9999999 -r ts-node/register './src/setResourcePriorities.test.ts'",
     "testSetEquippableList": "mocha --timeout 9999999 -r ts-node/register './src/setEquippableList.test.ts'",
     "testSendNft": "mocha --timeout 9999999 -r ts-node/register './src/sendNft.test.ts'",
