+++ tests/src/setCollectionProperty.test.ts
